Update to Example Code
I've placed commented lines in each section changed and listed reasons as to why I changed them. This code appears to work perfectly fine with my board and shield, even changing the month, days and years as expected.

If you have issues with this code, you potentially have an older chip being used in your RTC/SD Shield. Search for the manual on adafruit to determine which chip is being used in your shield.